Create PDF/A scripts for Short Write-Ups
Added a template and script to compile short write ups into PDF/A
format. The new tamplate is a copy of the ONLYONE template used for
PostScripts, but using the pdfx package. Renamed the runall script to a
more appropriate runps. The runall script now contains the logic for
running all PDF/A, PostScript and HTML generation.