Added missing media to anaphe
Add the missing images to the anaphe write-up. After spending some time
with Anaphe I have come to realize that it is one of the more complex
long write-ups so decided to leave it for later. PS and HTML versions of
the documentation do exist, so only the PDF/A would be useful to
generate.